@@ -1,23 +1,21 @@
 name: fish_redux
 description: Fish Redux is an assembled flutter application framework based on Redux state management.
-version: 0.3.7
+version: 0.4.0
 author: Alibaba Xianyu Team <tino.wjf@alibaba-inc.com>
 homepage: https://github.com/alibaba/fish-redux
 
 environment:
-    sdk: ">=2.2.0 <3.0.0"
+    sdk: '>=2.12.0 <3.0.0'
 dependencies:
     flutter:
         sdk: flutter
+    collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
     test: ^1.5.1
     flutter_test:
         sdk: flutter
-
-    test_widgets:
-        path: test/test_widgets
-
+        
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
 
