@@ -4,18 +4,18 @@ import '../redux/redux.dart';
 import '../redux_component/redux_component.dart';
 
 /// abstract for custom extends
-abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
-  final AdapterBuilder<T> _adapter;
+abstract class Adapter<T> extends Logic<T?> implements AbstractAdapter<T?> {
+  final AdapterBuilder<T?> _adapter;
 
-  AdapterBuilder<T> get protectedAdapter => _adapter;
+  AdapterBuilder<T?> get protectedAdapter => _adapter;
 
   Adapter({
-    @required AdapterBuilder<T> adapter,
-    Reducer<T> reducer,
-    ReducerFilter<T> filter,
-    Effect<T> effect,
-    Dependencies<T> dependencies,
-    @deprecated Object Function(T) key,
+    required AdapterBuilder<T?> adapter,
+    Reducer<T?>? reducer,
+    ReducerFilter<T?>? filter,
+    Effect<T?>? effect,
+    Dependencies<T>? dependencies,
+    @deprecated Object Function(T?)? key,
   })  : assert(adapter != null),
         assert(dependencies?.adapter == null,
             'Unexpected dependencies.list for Adapter.'),
@@ -30,25 +30,23 @@ abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
         );
 
   @override
-  ListAdapter buildAdapter(ContextSys<T> ctx) =>
-      ctx.enhancer
-          ?.adapterEnhance(protectedAdapter, this, ctx.store)
-          ?.call(ctx.state, ctx.dispatch, ctx) ??
-      protectedAdapter?.call(ctx.state, ctx.dispatch, ctx);
+  ListAdapter? buildAdapter(ContextSys<T?> ctx) => ctx.enhancer
+      .adapterEnhance(protectedAdapter, this, ctx.store)
+      .call(ctx.state, ctx.dispatch, ctx);
 
   @override
-  ContextSys<T> createContext(
-    Store<Object> store,
-    BuildContext buildContext,
-    Get<T> getState, {
-    @required Enhancer<Object> enhancer,
-    @required DispatchBus bus,
+  ContextSys<T?> createContext(
+    Store<Object?> store,
+    BuildContext? buildContext,
+    Get<T?> getState, {
+    void Function()? markNeedsBuild,
+    required Enhancer<Object?> enhancer,
+    required DispatchBus bus,
   }) {
-    assert(bus != null && enhancer != null);
-    return AdapterContext<T>(
+    return AdapterContext<T?>(
       logic: this,
       store: store,
-      buildContext: buildContext,
+      buildContext: buildContext!,
       getState: getState,
       bus: bus,
       enhancer: enhancer,
@@ -58,13 +56,13 @@ abstract class Adapter<T> extends Logic<T> implements AbstractAdapter<T> {
 
 class AdapterContext<T> extends LogicContext<T> {
   AdapterContext({
-    @required AbstractAdapter<T> logic,
-    @required Store<Object> store,
-    @required BuildContext buildContext,
-    @required Get<T> getState,
-    @required DispatchBus bus,
-    @required Enhancer<Object> enhancer,
-  })  : assert(bus != null && enhancer != null),
+    required AbstractAdapter<T> logic,
+    required Store<Object?> store,
+    required BuildContext buildContext,
+    required Get<T> getState,
+    required DispatchBus? bus,
+    required Enhancer<Object?> enhancer,
+  })   : assert(bus != null && enhancer != null),
         super(
           logic: logic,
           store: store,
@@ -75,8 +73,8 @@ class AdapterContext<T> extends LogicContext<T> {
         );
 
   @override
-  ListAdapter buildAdapter() {
-    final AbstractAdapter<T> curLogic = logic;
+  ListAdapter? buildAdapter() {
+    final AbstractAdapter<T?> curLogic = logic as AbstractAdapter<T?>;
     return curLogic.buildAdapter(this);
   }
 }
