@@ -9,8 +9,8 @@ import '../redux_component/redux_component.dart';
 /// }
 mixin PrivateReducerMixin<T> on Logic<T> {
   @override
-  Reducer<T> get protectedReducer {
-    final Reducer<T> superReducer = super.protectedReducer;
+  Reducer<T>? get protectedReducer {
+    final Reducer<T>? superReducer = super.protectedReducer;
     return superReducer != null
         ? (T state, Action action) {
             if (action is PrivateAction && action.target == state) {
@@ -22,7 +22,7 @@ mixin PrivateReducerMixin<T> on Logic<T> {
   }
 
   @override
-  Dispatch createDispatch(Dispatch effect, Dispatch next, Context<T> ctx) {
+  Dispatch createDispatch(Dispatch? effect, Dispatch next, Context<T> ctx) {
     final Dispatch superDispatch = super.createDispatch(effect, next, ctx);
     return (Action action) {
       if (action.type is! Lifecycle && action is! PrivateAction) {
@@ -38,7 +38,7 @@ mixin PrivateReducerMixin<T> on Logic<T> {
 }
 
 class PrivateAction extends Action {
-  final Object target;
+  final Object? target;
   PrivateAction(Object type, {dynamic payload, this.target})
       : super(type, payload: payload);
 
