@@ -13,36 +13,37 @@ import '../utils/utils.dart';
 /// }
 mixin VisibleChangeMixin<T> on AbstractAdapter<T> {
   @override
-  ListAdapter buildAdapter(ContextSys<T> ctx) {
-    return _wrapVisibleChange<T>(super.buildAdapter(ctx), ctx);
+  ListAdapter? buildAdapter(ContextSys<T> ctx) {
+    return _wrapVisibleChange<T>(
+        super.buildAdapter(ctx), ctx as LogicContext<T>);
   }
 }
 
 class _VisibleChangeState extends State<_VisibleChangeWidget> {
   @override
   Widget build(BuildContext context) =>
-      widget.itemBuilder(context, widget.index);
+      widget.itemBuilder!(context, widget.index!);
 
   @override
   void initState() {
     super.initState();
-    widget.dispatch(LifecycleCreator.appear(widget.index));
+    widget.dispatch!(LifecycleCreator.appear(widget.index));
   }
 
   @override
   void dispose() {
-    widget.dispatch(LifecycleCreator.disappear(widget.index));
+    widget.dispatch!(LifecycleCreator.disappear(widget.index));
     super.dispose();
   }
 }
 
 class _VisibleChangeWidget extends StatefulWidget {
-  final IndexedWidgetBuilder itemBuilder;
-  final int index;
-  final Dispatch dispatch;
+  final IndexedWidgetBuilder? itemBuilder;
+  final int? index;
+  final Dispatch? dispatch;
 
   const _VisibleChangeWidget({
-    Key key,
+    Key? key,
     this.itemBuilder,
     this.index,
     this.dispatch,
@@ -52,12 +53,12 @@ class _VisibleChangeWidget extends StatefulWidget {
   State<StatefulWidget> createState() => _VisibleChangeState();
 }
 
-ListAdapter _wrapVisibleChange<T>(
-  ListAdapter listAdapter,
+ListAdapter? _wrapVisibleChange<T>(
+  ListAdapter? listAdapter,
   LogicContext<T> ctx,
 ) {
-  final _VisibleChangeDispatch onChange =
-      (ctx.extra['\$visible'] ??= _VisibleChangeDispatch(ctx.dispatch));
+  final _VisibleChangeDispatch onChange = (ctx.extra['\$visible'] ??=
+      _VisibleChangeDispatch(ctx.dispatch)) as _VisibleChangeDispatch;
 
   return listAdapter == null
       ? null
