@@ -6,7 +6,7 @@ import '../aop.dart';
 ApplyLikeEnhancer delay(int millis) {
   return (dynamic Function(List<dynamic>) functor) {
     return (List<dynamic> positionalArguments,
-        [Map<Symbol, dynamic> namedArguments]) async {
+        [Map<Symbol, dynamic>? namedArguments]) async {
       await Future<void>.delayed(Duration(milliseconds: millis));
       return functor(positionalArguments);
     };
