@@ -10,11 +10,11 @@ ApplyLikeEnhancer performanceAOP(String tag) {
   return isDebug()
       ? (dynamic Function(List<dynamic>) functor) {
           return (List<dynamic> positionalArguments,
-              [Map<Symbol, dynamic> namedArguments]) {
+              [Map<Symbol, dynamic>? namedArguments]) {
             final int marked = DateTime.now().microsecondsSinceEpoch;
             final Object result = functor(positionalArguments);
             if (result is Future) {
-              result.then((Object r) {
+              result.then<dynamic>((dynamic r) {
                 print(
                     '$tag performance <Future>: ${_microSecsSinceEpoch() - marked}');
                 return r;
