@@ -8,7 +8,7 @@ ApplyLikeEnhancer throttle(int millis) {
   return (dynamic Function(List<dynamic>) functor) {
     int last = 0;
     return (List<dynamic> positionalArguments,
-        [Map<Symbol, dynamic> namedArguments]) {
+        [Map<Symbol, dynamic>? namedArguments]) {
       final int now = _microSecsSinceEpoch();
       final int elapsed = now - last;
       if (elapsed >= millis) {
