@@ -6,11 +6,11 @@ ApplyLikeEnhancer waitUntil() {
   return (dynamic Function(List<dynamic>) functor) {
     bool isLocked = false;
     return (List<dynamic> positionalArguments,
-        [Map<Symbol, dynamic> namedArguments]) {
+        [Map<Symbol, dynamic>? namedArguments]) {
       if (isLocked) {
         return null;
       } else {
-        final Object result = functor(positionalArguments);
+        final Object? result = functor(positionalArguments);
         if (result is Future) {
           isLocked = true;
           return result.whenComplete(() {
