@@ -34,7 +34,7 @@ enum Lifecycle {
 class LifecycleCreator {
   static Action initState() => const Action(Lifecycle.initState);
 
-  static Action build(String name) => Action(Lifecycle.build, payload: name);
+  static Action build(String? name) => Action(Lifecycle.build, payload: name);
 
   static Action reassemble() => const Action(Lifecycle.reassemble);
 
@@ -49,9 +49,9 @@ class LifecycleCreator {
 
   static Action deactivate() => const Action(Lifecycle.deactivate);
 
-  static Action appear(int index) => Action(Lifecycle.appear, payload: index);
+  static Action appear(int? index) => Action(Lifecycle.appear, payload: index);
 
-  static Action disappear(int index) =>
+  static Action disappear(int? index) =>
       Action(Lifecycle.disappear, payload: index);
 
   static Action didChangeAppLifecycleState(AppLifecycleState state) =>
