@@ -6,13 +6,13 @@ import '../../utils/utils.dart';
 
 ViewMiddleware<T> safetyView<T>(
     {Widget Function(dynamic, StackTrace,
-            {AbstractComponent<dynamic> component, Store<T> store})
+            {AbstractComponent<dynamic>? component, Store<T>? store})?
         onError}) {
-  return (AbstractComponent<dynamic> component, Store<T> store) {
+  return (AbstractComponent<dynamic> component, Store<T>? store) {
     return (ViewBuilder<dynamic> next) {
       return isDebug()
           ? next
-          : (dynamic state, Dispatch dispatch, ViewService viewService) {
+          : (dynamic state, Dispatch dispatch, ViewService? viewService) {
               try {
                 return next(state, dispatch, viewService);
               } catch (e, stackTrace) {
