@@ -3,7 +3,7 @@ import 'dart:core';
 /// Util for collections.
 class Collections {
   /// Wrap List.reduce with a check list is null or empty.
-  static E reduce<E>(List<E> list, E combine(E e0, E e1)) =>
+  static E? reduce<E>(List<E> list, E combine(E e0, E e1)) =>
       (list == null || list.isEmpty) ? null : list.reduce(combine);
 
   /// Wrap List.fold with a check list is null or empty.
@@ -16,7 +16,7 @@ class Collections {
   /// List<String> b = ['1', '2', '3'];
   /// List<List<String>> list = [a, b] // [[a, b, c], [1, 2, 3]]
   /// List<String> listFlatten = Collections.flatten(list) // [a, b, c, 1, 2, 3]
-  static List<E> flatten<E>(List<List<E>> lists) => reduce(lists, merge);
+  static List<E>? flatten<E>(List<List<E>> lists) => reduce(lists, merge);
 
   /// Merge two Iterable
   /// List<String> a = ['a', 'b', 'c'];
@@ -48,7 +48,7 @@ class Collections {
   /// print(list)                       // [1, 2, null, 3, null]
   /// print(Collections.compact(list)); // [1, 2, 3]
   static List<T> compact<T>(Iterable<T> list, {bool growable = true}) =>
-      list?.where((T e) => e != null)?.toList(growable: growable);
+      list.where((T e) => e != null).toList(growable: growable);
 
   /// Check if an Object is Empty.
   static bool isEmpty(Object value) {
